requirements.txt
pillow==11.1.0
requests==2.32.4
streamlit==1.51.0
streamlit-lottie==0.0.5

